GUI: fix incorrect reading of rotation axis info for turbomachinery.
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@7724 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ src/gui/cs_gui.c
@@ -2075,7 +2075,7 @@ _c_head_losses(const char* zone_id, const char* c)
  *   name        <--  name of the property
  *----------------------------------------------------------------------------*/
 
-static int
+static double
 _rotor_option(int          rotor_id,
               const char  *name)
 {
